app/requirements.txt
psycopg2
pymongo
dnspython
python-dotenv
fastapi_jwt_auth
bcrypt
uvicorn